Develop

+++ Web_App/Rest/Cart/Model/CartModelResponse.cs
@@ -6,9 +6,9 @@ namespace Web_App.Rest.Cart.Model
     {
         public int TowarID { get; set; }
         public string TowarName { get; set; }
-        public float TowarPrice { get; set; }
+        public double TowarPrice { get; set; }
         public int Count { get; set; }  
-        public float SumPrice { get; set; }
+        public double SumPrice { get; set; }
         public byte[] Image { get; set; }
