+++ Web_App/Rest/Favorite/Models/FavoritePageResponseModel.cs
@@ -4,7 +4,7 @@
     {
         public int ProductID { get; set; }
         public string ProductName { get; set; }
-        public float Cost { get; set; }
+        public double Cost { get; set; }
         public string Description { get; set; }
         public byte[] ImageUrl { get; set; }
